requirements.txt
numpy
pandas
matplotlib
Pillow
torch
torchvision
opencv-python
streamlit
